requirements.txt
asgiref==3.6.0
cloudinary==1.31.0
dj-database-url==1.2.0
dj3-cloudinary-storage==0.0.6
Django==3.2.16
django-entangled==0.5.3
django-json-widget==1.1.1
django-prettyjson==0.4.1
future==0.18.3
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
standardjson==0.3.1
